fix(g-abstract, g-canvas, g-svg): elevate events methods to g-abstract
