+++ packages/g-canvas/src/canvas.ts
@@ -1,7 +1,6 @@
 import { AbstractCanvas } from '@antv/g-base';
 import { ChangeType } from '@antv/g-base/lib/types';
 import { IElement } from './interfaces';
-import EventController from '@antv/g-base/lib/event/event-contoller';
 import * as Shape from './shape';
 import Group from './group';
 import { applyAttrsToContext, drawChildren, getMergedRegion } from './util/draw';
@@ -60,21 +59,6 @@ class Canvas extends AbstractCanvas {
       maxX: element.width,
       maxY: element.height,
     };
-  }
-
-  // 复写处理事件
-  initEvents() {
-    const eventController = new EventController({
-      canvas: this,
-    });
-    eventController.init();
-    this.set('eventController', eventController);
-  }
-
-  // 覆盖基类中的 clearEvents 方法
-  clearEvents() {
-    const eventController = this.get('eventController');
-    eventController.destroy();
   }
 
   // 复写基类的方法生成标签
@@ -208,12 +192,6 @@ class Canvas extends AbstractCanvas {
   }
 
   skipDraw() {}
-
-  destroy() {
-    const eventController = this.get('eventController');
-    eventController.destroy();
-    super.destroy();
-  }
 }
 
 export default Canvas;
